Editorial: Rename some script-related aliases (#2353)
In ParseScript:
ParseText is invoked with |Script| as the goal,
so the result (if it isn't errors) is an instance of |Script|,
so change the alias `_body_` to `_script_`.

And note that the Script Record's [[ECMAScriptCode]] field
is also an instance of |Script|.

In ScriptEvaluation:
Since _scriptRecord_.[[ECMAScriptCode]] is an instance of |Script|,
change the alias `_scriptBody_` to `_script_`.