Editorial: Define `Evaluation` for BindingIdentifier
PR #614 defined (in Annex B) an additional form for ForInOfStatement.
Performing ForInOfLoopEvaluation on it ends with a call to ForIn/OfBodyEvaluation,
with |BindingIdentifier| as first arg, passed to the _lhs_ param.
ForIn/OfBodyEvaluation (at step 6.g.i.1) evaluates this param,
but there's no definition of Evaluation for |BindingIdentifier|.
This PR adds the necessary definition.

(The new definition basically just duplicates the one for ForBinding.
In fact, if this part of Annex B is merged into the main spec,
the definition for ForBinding can be dropped,
since it can just chain to this new definition.)