Reorganize
Looking at this again...

(1)
I don't think CreateBuiltinFunction should be getting into
the details of what internal slots are appropriate.
That should all be covered elsewhere (i.e., in 10.3).

(2)
I don't think 10.3 needs the table.

Note that it lists [[Realm]], and the following para
adds on the table for ordinary function objects,
which also lists [[Realm]].

Note also that there's an earlier paragraph
that lists the slots for an exotic built-in,
but leaves out [[InitialName]],
so we're giving mixed messages.

Instead, I think 10.3 can cover it in 2 paragraphs,
that concisely and (fairly) clearly
specify all the internal slots for a built-in function,
depending on how it's implemented.