Editorial: Delete redundant xrefs from Well-Known Intrinsics table
In the `Well-Known Intrinsic Objects` table
(and the `Additional Well-known Intrinsic Objects` table),
under "ECMAScript Language Association",
most entries have an `<emu-xref>`,
but in most cases it links to the same place
as the "Intrinsic Name" will be auto-linked to.
Delete all these redundant xrefs.

(The remaining 7 cases are all for %FooIteratorPrototype%,
where the auto-link goes to the clause that 'declares' the intrinsic,
but the "Language Association" xref goes to the clause
that deals with Foo iterator objects in general.)