Transfer control: Replace 3 "Return" steps
In GeneratorStart, AsyncGeneratorStart, and AsyncBlockStart,
the Abstract Closure ends with a `Return`.
This is odd, because where would it return to?
Normally, a `Return` would return to the operation that invoked the Abstract Closure,
but in this case, the invocation mechanism is unusual, and in general,
the operation that caused it is no longer waiting for a result.

Instead, the `Return` is presumably supposed to send control back to
the latest next-side `Transfer control` step.
So make that explicit.