Resume+Suspend: Eliminate `Resume` steps
The only remaining `Resume` steps are the 4 non-blocking Resumes
in ScriptEvaluation, ExecuteModule, Evaluate, and PerformEval.

Since there's no definition for "resuming" an execution context,
and since it doesn't come up any other time that an EC
is removed from the EC stack,
I'm concluding that it has no normative effect.

So remove the remaining `Resume` steps.