Execution Context section: Make the EC Stack subsection
Collect stuff about the execution context stack,
move it to a new subsection, and then rework it.

Note that there a couple of problems in the status quo:
"A new execution context is created whenever control is transferred from the executable code associated with the currently running execution context to executable code that is not associated with that execution context."
That became false when generators were added.

"Transition of the running execution context status among execution contexts usually occurs in stack-like last-in/first-out manner. However, some ECMAScript features require non-LIFO transitions of the running execution context."
No, it's always stack-like.